backend/pdf-label/requirements.txt
reportlab==4.0.7
psycopg2-binary==2.9.9
